// src/app/core/models/application.model.ts
export interface Application {
  id?: number;              // auto-generated by JSON Server
  userId: string;           // ID of the logged user
  offerId: number | string; // ID of the job offer
  apiSource: string;        // API source (adzuna, jooble...)
  title: string;            // job title
  company: string;          // company name
  location: string;         // job location
  url: string;              // link to the job offer
  status: 'en_attente' | 'accepte' | 'refuse'; // application status
  notes?: string;           // optional notes
  dateAdded: string;        // ISO date
}